Asyncronous compuation, progress bar 	modified:   metascript.m

+++ metascript.m
@@ -1,6 +1,6 @@
 logDir = '../ExperimentData/';
 maxIter = 20;
-parpool(4);
+workers_num = 4;
 
 if ~ 7 == exist(logDir, 'dir')
     mkdir(logDir);
@@ -160,6 +160,7 @@ fml9.preds = [fml3.preds, pred];
 fml9.stopTime = 100;
 
 formulas = {fml1, fml2, fml3, fml4, fml5, fml6, fml7, fml8, fml9 };
+%formulas = {fml6};
 
 configs = { };
 for k = 1:size(formulas, 2)
@@ -174,36 +175,60 @@ for k = 1:size(formulas, 2)
 end
  load_system(mdl);
  
- spmd
-    % Setup tempdir and cd into it
-    currDir = pwd;
-    addpath(currDir);
-     P = py.sys.path;
-    insert(P,int32(0),pwd);
-    tmpDir = tempname;
-    mkdir(tmpDir);
-    cd(tmpDir);
-    % Load the model on the worker
-    load_system(mdl);
+ if workers_num > 1
+     parpool(workers_num);    
+     spmd
+        % Setup tempdir and cd into it
+        currDir = pwd;
+        addpath(currDir);
+         P = py.sys.path;
+        insert(P,int32(0),pwd);
+        tmpDir = tempname;
+        mkdir(tmpDir);
+        cd(tmpDir);
+        % Load the model on the worker
+        load_system(mdl);
+     end
+
+     results = cell(size(configs ,2));
+     
+     for idx = 1:size(configs, 2)
+        F(idx) = parfeval(@falsify,1,configs{idx});
+     end
+     % Build a waitbar to track progress
+     h = waitbar(0,'Waiting for FevalFutures to complete...');
+     for idx = 1:size(configs, 2)
+        [completedIdx,thisResult] = fetchNext(F);
+        % store the result
+        results{completedIdx} = thisResult;
+        % update waitbar
+        waitbar(idx/size(configs, 2),h,sprintf('Latest result: %d',thisResult));
+     end
+     delete(h)
+     
+     spmd
+     cd(currDir);
+     rmdir(tmpDir,'s');
+     rmpath(currDir);
+     close_system(mdl, 0);
+     end
+    
+     delete(gcp('nocreate'));
+ else
+     h = waitbar(0,'Please wait...');
+     results = cell(size(configs));
+     for i = 1:size(configs, 2)
+        config = configs{i};
+        results{i} = experiment_rl(config);
+        waitbar(i / size(configs))
+     end
+     close(h)
  end
- 
- results = cell(size(configs));
- parfor i = 1:size(configs, 2)
-    config = configs{i};
-    results{i} = experiment_rl(config);
- end
- 
+   
  save(logFile, 'configs', 'results');
  
- spmd
-    cd(currDir);
-    rmdir(tmpDir,'s');
-    rmpath(currDir);
-    close_system(mdl, 0);
- end
 
 close_system(mdl, 0);
-delete(gcp('nocreate'));
 
 function [iter_result] = experiment_rl(config)
     iter_result = cell([1, config.maxIter]);
